Ajustando dropdown

+++ app/models.py
@@ -12,8 +12,8 @@ class Amortecedor (models.Model):
 
 	#amortecedor_id = models.AutoField(primary_key=True, verbose_name='ID do Amortecedor')
 	amortecedor_codigo = models.CharField(max_length=45, primary_key=True, verbose_name='Código') 
-	amortecedor_diametro_externo = models.IntegerField(verbose_name='Tamanho do Diâmetro Externo')
-	amortecedor_curso = models.IntegerField(verbose_name='Tamanho do Curso')
+	amortecedor_diametro_externo = models.FloatField(verbose_name='Tamanho do Diâmetro Externo')
+	amortecedor_curso = models.FloatField(verbose_name='Tamanho do Curso')
 
 	def __repr__(self):
 		return self.amortecedor_codigo
